+++ pyoxidizer/src/py_packaging/distribution.rs
@@ -10,7 +10,6 @@ use {
     super::binary::PythonBinaryBuilder,
     super::bytecode::BytecodeCompiler,
     super::config::EmbeddedPythonConfig,
-    super::libpython::ImportlibBytecode,
     super::resource::{PythonModuleSource, PythonPackageResource, PythonResource},
     super::resources_policy::PythonResourcesPolicy,
     super::standalone_distribution::{DistributionExtensionModule, StandaloneDistribution},
@@ -137,15 +136,6 @@ pub trait PythonDistribution {
 
     /// Create a `BytecodeCompiler` from this instance.
     fn create_bytecode_compiler(&self) -> Result<BytecodeCompiler>;
-
-    /// Resolve `importlib` bytecode.
-    ///
-    /// The returned struct holds Python bytecode which will be turned into
-    /// a frozen module. This bytecode registers the PyOxidizer memory importer.
-    ///
-    /// The bytecode should be compiled from modified sources of the
-    /// corresponding Python modules.
-    fn resolve_importlib_bytecode(&self) -> Result<ImportlibBytecode>;
 
     /// Obtain a `PythonBinaryBuilder` for constructing an executable embedding Python.
     ///